image-android-sparse: make the CRC optional
Recent Android fastboot tools don't support sparse images containing the
CRC at the end of the image and instead fail with an coredump. Checking
the fastboot code base there are no promising commits that this will
change in the near future:

| commit 94d05063ba7d2a2d614733d179c4bce94cead362
| Author: Hridya Valsaraju <hridya@google.com>
| Date:   Tue Oct 2 10:08:38 2018 -0700
|
|     Remove tests for sparse images with CRC from fuzzy_fastboot
|
|     Sparse images with CRC are currently not supported by the
|     Android build system.
|
|     Test: ./fuzzy_fastboot --gtest_filter=Conformance.*
|     Change-Id: I6a73fb0dc5d4c77a3d1a0384c137789eb06fa956

Therefore make the crc optional and turn it of by default to have an
better user experience. This is also what img2simg does except for the
fact that they don't support adding the crc at all.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>